requirements-devel.txt
pytest
python-docx
sphinx_rtd_theme
Pygments
